refactor: move comparison state to compare/
